refactor predictor
