+++ multimodal/src/autogluon/multimodal/learners/__init__.py
@@ -1,0 +1,3 @@
+from .base import BaseLearner
+from .object_detection import ObjectDetectionLearner
+from .ner import NERLearner
